% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark_functions.r
\name{calcul_prsim_vers_statistiques_sommaires_temiscamingue}
\alias{calcul_prsim_vers_statistiques_sommaires_temiscamingue}
\title{calcul_prsim_vers_statistiques_sommaires_temiscamingue}
\usage{
calcul_prsim_vers_statistiques_sommaires_temiscamingue(path)
}
\arguments{
\item{path}{}
}
\value{

}
\description{
calcul_prsim_vers_statistiques_sommaires_temiscamingue
}
