% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark_functions.r
\name{ecriture_prsim_vers_csv}
\alias{ecriture_prsim_vers_csv}
\title{ecriture_prsim_vers_csv}
\usage{
ecriture_prsim_vers_csv(path)
}
\arguments{
\item{path}{}
}
\value{

}
\description{
ecriture_prsim_vers_csv
}
\examples{
mainDir<-'/media/tito/TIIGE/PRSIM/0.9995/bv_csv_hecressim'
}
